requirements.txt
easydict==1.11
numpy==1.23.4
pytorch_lightning==2.1.2
PyYAML==6.0.1
torch==2.0.1+cu118
torch_scatter==2.1.1+pt20cu118
transformers==4.32.1